* 'master' of https://gitee.com/chendawei8/web_webview: (52 commits)
  ffi_fix_code_check
  update ohos_interface/ohos_glue/ohos_adapter/include/ark_ohos_native_buffer_adapter.h.
  WebviewHasImageCallback::UvAfterWorkCbAsync存在内存泄漏
  update test/unittest/ohos_adapter/native_buffer_adapter_test/native_buffer_adapter_test.cpp.
  test
  fix
  rm test
  rm unused code
  update ohos_interface/ohos_glue/ohos_adapter/include/ark_graphic_adapter.h.
  notes
  NWebCreateNativeMediaPlayerCallbackImpl::OnCreate存在内存泄露问题
  update ohos_interface/ohos_glue/ohos_adapter/include/ark_ohos_native_buffer_adapter.h.
  update ohos_interface/ohos_glue/ohos_adapter/include/ark_ohos_adapter_helper.h.
  web ax filter
  update ohos_interface/ohos_glue/ohos_adapter/include/ark_graphic_adapter.h.
  add tdd .
  SensorCallback Update
  update ArkWebCore.hap in master to 0923
  add tdd.
  update ohos_interface/ohos_glue/ohos_adapter/include/ark_ohos_adapter_helper.h.
  ...

+++ interfaces/kits/cj/src/webview_controller_impl.cpp
@@ -1002,12 +1002,12 @@ namespace OHOS::Webview {
     
     int32_t WebviewControllerImpl::CustomizeSchemesArrayDataHandler(CArrScheme schemes)
     {
-        uint32_t arrayLength = schemes.size;
+        int64_t arrayLength = schemes.size;
         if (arrayLength > MAX_CUSTOM_SCHEME_SIZE) {
             return NWebError::PARAM_CHECK_ERROR;
         }
         std::vector<Scheme> schemeVector;
-        for (uint32_t i = 0; i < arrayLength; ++i) {
+        for (int64_t i = 0; i < arrayLength; ++i) {
             Scheme scheme;
             bool result = SetCustomizeScheme(schemes.cScheme[i], scheme);
             if (!result) {
